requirements.txt
beautifulsoup4==4.12.3
googlesearch_python==1.2.5
requests
torch
